requirements.txt
pandas~=1.5.2
transformers~=4.23.1
fastparquet~=2023.1.0
pyarrow~=11.0.0
html5lib~=1.1
torch~=1.11.0
elasticsearch-dsl~=7.4.1
Django~=4.1.7
beautifulsoup4~=4.11.2
faiss-cpu~=1.7.3
numpy~=1.24.2
lxml~=4.9.2